refactor: Regroup component related to Avatar together
