+++ packages/cozy-sharing/src/components/Avatar/AvatarPlusX.jsx
