+++ packages/cozy-sharing/src/components/Avatar/AvatarPlusX.stories.jsx
