+++ packages/cozy-sharing/src/components/Avatar/RecipientAvatar.jsx
