+++ packages/cozy-sharing/src/components/Avatar/RecipientAvatar.stories.jsx
