+++ packages/cozy-sharing/src/components/Avatar/RecipientsAvatars.jsx
