+++ packages/cozy-sharing/src/components/Avatar/RecipientsAvatars.spec.jsx
