+++ packages/cozy-sharing/src/components/Avatar/RecipientsAvatars.stories.jsx
